spell-checker/pyproject.toml
[tool.poetry]
name = "spell-checker"
version = "0.1.0"
description = ""
authors = ["Golovin Pavel <gopavel0@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^19.3.0"
tqdm = "^4.45.0"
numpy = "^1.18.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
